requirements.txt
streamlit==1.31.1
plotly==5.18.0
pandas==2.2.0
requests==2.31.0
pyserial==3.5
mpremote==1.0.0
streamlit-folium==0.15.1
numpy==1.26.3 